fix: refactor hook
